Refactor controllers requiring auth to be under /admin
