+++ src/Controller/QrController.php
@@ -8,26 +8,27 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Uid\UuidV7;
 
-final class RedirectController extends AbstractController
+final class QrController extends AbstractController
 {
     public function __construct(
         private readonly QrRepository $qrRepository,
     ) {
     }
 
-    #[Route('/redirect/{uuid}', name: 'app_redirect')]
+    #[Route('/qr/{uuid}', name: 'app_qr_index')]
     public function index(string $uuid, UrlRepository $urlRepository): Response
     {
         // Find the QR entity by UUID
+        $uuid = UuidV7::fromString($uuid);
         $qr = $this->qrRepository->findOneBy(['uuid' => $uuid]);
 
         if (!$qr) {
             throw $this->createNotFoundException('QR code not found');
         }
 
-        // Retrieve URLs associated with the QR entity
-        $urls = $urlRepository->findBy(['qr' => $qr->getId()]);
+        $urls = $qr->getUrls();
 
         // @TODO: Add what happens if a qr has multiple urls with certain modes.
 
@@ -36,6 +37,7 @@ final class RedirectController extends AbstractController
         }
 
         // Redirect to the first URL
+        // @TODO enable "kiosk mode" for codes with multiple urls
         return new RedirectResponse((string) $urls[0]);
     }
 }
